requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.13.1
anyio==3.7.1
arxiv==1.4.8
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.4.0
autopep8==2.0.2
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.22.1
black==23.7.0
blinker==1.8.0
blis==0.7.10
bs4==0.0.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
chainlit==0.5.2
chardet==5.2.0
charset-normalizer==3.2.0
chromadb==0.3.26
click==8.1.7
clickhouse-connect==0.6.8
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.1
contourpy==1.1.0
coolsms-python-sdk==2.0.3
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.9
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
duckdb==0.8.1
emoji==2.8.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.97.0
fastapi-socketio==0.0.10
feedparser==6.0.10
filetype==1.2.0
Flask==3.0.3
Flask-DebugToolbar==0.15.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fluent-logger==0.10.0
fonttools==4.42.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-search-results==2.4.2
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
gTTS==2.5.1
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
ipython==8.14.0
isort==5.12.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
langchain==0.0.261
langcodes==3.3.0
langsmith==0.0.24
lark==1.1.7
lazy-object-proxy==1.9.0
libmagic==1.0
loguru==0.7.0
lxml==4.9.3
lz4==4.3.2
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pathy==0.10.2
pdf2image==1.17.0
pexpect==4.9.0
phonenumbers==8.13.36
pickleshare==0.7.5
Pillow==10.0.0
pinecone-client==2.2.2
platformdirs==3.10.0
playwright==1.37.0
posthog==3.0.2
preshed==3.0.8
prisma==0.9.1
prompt-toolkit==3.0.39
protobuf==4.24.0
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pycodestyle==2.11.0
pycparser==2.21
pydantic==1.10.12
pyee==9.0.4
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.5
PyMuPDF==1.22.5
PyMySQL==1.1.0
pyparsing==3.0.9
pypdf==3.15.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-graphql-client==0.4.3
python-magic==0.4.27
python-socketio==5.8.0
pytube==15.0.0
pytz==2023.3
PyYAML==6.0.1
rank-bm25==0.2.2
redis==4.6.0
regex==2023.8.8
reportlab==4.2.0
requests==2.28.2
rich==13.5.2
ruff==0.0.285
semver==3.0.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.20
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
steamship==2.17.24
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.3
thinc==8.1.12
tiktoken==0.3.3
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.9.3
uptrace==1.19.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==3.0.2
wikipedia==1.4.0
win32-setctime==1.1.0
wrapt==1.15.0
yapf==0.40.1
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.16.2
zstandard==0.21.0
